[Silver V] Title: 귀찮음, Time: 444 ms, Memory: 57608 KB, Score: 30 point -BaekjoonHub

+++ 백준/Silver/16208. 귀찮음/귀찮음.java
@@ -18,11 +18,11 @@ public class Main {
             sum += arr[i];
         }
 
-        int result = 0;
+        long result = 0;
         Arrays.sort(arr);
 
         for (int i = 0; i < N - 1; i++) {
-            result += arr[i] * (sum - arr[i]);
+            result += (long) arr[i] * (sum - arr[i]);
             sum -= arr[i];
         }
         System.out.println(result);
